Remove old Hosting overview codes and move the used files to the new folder.
