--- client/a8c-for-agencies/sections/marketplace/hosting-overview/hosting-v2/enterprise-agency-hosting/style.scss
@@ -1,73 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-
-.enterprise-agency-hosting__top-container {
-	display: grid;
-	grid-template-columns: 1fr;
-	gap: 16px;
-	padding: 0 16px 48px;
-
-	@include break-medium {
-		max-width: 1500px;
-		margin-inline: auto;
-		padding-inline: 40px;
-		gap: 24px;
-	}
-
-	@include break-xlarge {
-		grid-template-columns: 1fr 1fr;
-	}
-
-	.enterprise-agency-hosting__details {
-		border: 1px solid var(--color-accent-5);
-		border-radius: 4px;
-		padding: 24px;
-		gap: 24px;
-		display: flex;
-		flex-direction: column;
-		justify-content: space-between;
-
-		&:nth-child(2) {
-			background-color: #f5f2f1;
-			justify-content: start;
-		}
-	}
-
-	.enterprise-agency-hosting__details-heading {
-		@include a4a-font-heading-xl;
-	}
-
-	.enterprise-agency-hosting__details-subheading {
-		padding-block-end: 24px;
-	}
-
-	.enterprise-agency-hosting__logos {
-		padding-block: 24px;
-		display: grid;
-		grid-template-columns: repeat(2, 1fr);
-		gap: 24px;
-		justify-items: center;
-
-		@include break-medium {
-			grid-template-columns: repeat(3, 1fr);
-		}
-
-		@include break-wide {
-			grid-template-columns: repeat(4, 1fr);
-		}
-
-		svg,
-		img {
-			width: 120px;
-			height: 40px;
-
-			@include break-xlarge {
-				width: auto;
-			}
-
-			@include break-huge {
-				width: 120px;
-			}
-		}
-	}
-}
