--- client/a8c-for-agencies/sections/marketplace/hosting-overview/hosting-v2/premier-agency-hosting/style.scss
@@ -1,86 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-
-.premier-agency-hosting .pressable-overview-plan-selection {
-	padding: 0 16px;
-	box-sizing: border-box;
-
-	.pressable-overview-plan-selection__filter-owned-plan {
-		margin-block-start: 0;
-
-		.badge {
-			font-size: rem(12px);
-
-			@include break-mobile {
-				font-size: rem(14px);
-			}
-		}
-	}
-
-	.pressable-overview-plan-selection__details-card {
-		border: none;
-		padding: 0;
-	}
-
-	.pressable-overview-plan-selection__filter-type {
-		padding-block-end: 32px;
-	}
-
-	.pressable-overview-plan-selection__filter-buttons {
-		display: flex;
-	}
-
-	@include break-large {
-		max-width: 1000px;
-		width: 100%;
-
-		.pressable-overview-plan-selection__details-card {
-			border: 1px solid var(--color-neutral-5);
-			padding: 24px;
-		}
-	}
-}
-
-.pressable-overview-plan-existing {
-	background-image: linear-gradient(0deg, var(--color-primary-0) 0%, #fff 100%);
-	padding: 0 16px 48px;
-	margin-bottom: 32px;
-
-	@include break-large {
-		padding: 0 32px 48px;
-	}
-
-	.pressable-overview-plan-existing__details-card{
-		flex: 1;
-		border: 1px solid var(--color-neutral-5);
-		border-radius: 4px;
-		padding: 16px;
-		justify-content: space-between;
-		max-width: 1000px;
-		margin: 0 auto;
-		background-color: #fff;
-
-		@include break-large {
-			padding: 24px;
-		}
-	}
-
-	.pressable-overview-plan-existing__owned-plan {
-		font-size: 0.75rem;
-		color: var(--color-neutral-60);
-	}
-	.pressable-overview-plan-existing__name {
-		font-size: 1.50rem;
-		font-weight: bold;
-		color: var(--color-accent-100);
-	}
-
-	.pressable-overview-plan-existing__header {
-		margin-bottom: 16px;
-		text-align: center;
-
-		@include break-large {
-			text-align: left;
-		}
-	}
-}
