--- client/a8c-for-agencies/sections/marketplace/hosting-overview/hosting-v2/standard-agency-hosting/style.scss
@@ -1,24 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-
-.standard-agency-hosting__plan-selector-container {
-	padding: 0 16px 48px;
-
-	@include break-medium {
-		max-width: 1500px;
-		margin-inline: auto;
-		padding-inline: 40px;
-	}
-}
-
-.standard-agency-hosting__banner-heading {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	gap: 8px;
-	text-align: center;
-}
-
-.standard-agency-hosting__banner-description {
-	@include a4a-font-body-md;
-}
