--- client/a8c-for-agencies/sections/marketplace/hosting-overview/hosting-v2/standard-agency-hosting/wpcom-plan-selector/style.scss
@@ -1,170 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-
-.wpcom-plan-selector {
-	display: flex;
-	flex-direction: column;
-	gap: 32px;
-
-	&.is-slider-hidden {
-		gap: 24px;
-	}
-}
-
-.wpcom-plan-selector__slider-container {
-	max-width: 900px;
-	width: 100%;
-	margin: 0 auto;
-
-	@include break-huge {
-		.a4a-slider {
-			margin-inline-start: -93px;
-		}
-	}
-}
-
-.wpcom-plan-selector__card {
-	display: flex;
-	flex-direction: column;
-	gap: 24px;
-
-	@include break-medium {
-		border: 1px solid var(--color-neutral-5);
-		border-radius: 4px;
-		padding: 32px;
-	}
-
-	@include break-xlarge {
-		flex-direction: row;
-		gap: 108px;
-	}
-
-}
-
-.wpcom-plan-selector__plan-name {
-	@include a4a-font-heading-xl;
-}
-
-.wpcom-plan-selector__price-actual-value,
-.wpcom-plan-selector__price-original-value {
-	@include a4a-font-heading-lg;
-	color: var(--color-text);
-}
-
-.wpcom-plan-selector__price-original-value {
-	display: inline-block;
-	text-decoration: line-through;
-	margin-inline: 8px;
-}
-
-.wpcom-plan-selector__price-discount {
-	@include a4a-font-body-md;
-	color: var(--color-success);
-	text-wrap: nowrap;
-}
-
-.wpcom-plan-selector__price-interval {
-	@include a4a-font-body-sm;
-}
-
-.wpcom-plan-selector__cta {
-	display: flex;
-	flex-direction: column;
-	align-items: stretch;
-	gap: 16px;
-	margin-block-start: 24px;
-}
-
-.wpcom-plan-selector__cta-label {
-	@include a4a-font-body-md;
-}
-
-.wpcom-plan-selector__features {
-	display: flex;
-	flex-direction: column;
-	justify-content: space-evenly;
-	flex-grow: 1;
-
-	@include break-wide {
-		flex-direction: row;
-		gap: 48px;
-	}
-}
-
-.wpcom-plan-selector__owned-plan {
-	@include a4a-font-heading-md;
-
-	display: inline-block;
-	border: 1px solid var(--color-neutral-5);
-	background-color: var(--color-neutral-0);
-	padding: 4px 16px;
-	/* stylelint-disable-next-line scales/radii */
-	border-radius: 16px;
-	margin-block-end: 8px;
-}
-
-.wpcom-plan-selector__cta-component {
-	display: flex;
-	flex-direction: column-reverse;
-	align-items: flex-start;
-	gap: 16px;
-
-	@include break-huge {
-		flex-direction: row;
-	}
-}
-
-.wpcom-plan-selector__cta-button {
-	min-width: 260px;
-	width: 100%;
-}
-
-.wpcom-plan-selector__details.is-placeholder {
-	display: flex;
-	flex-direction: column;
-	align-items: flex-start;
-
-	.wpcom-plan-selector__owned-plan,
-	.wpcom-plan-selector__plan-name,
-	.wpcom-plan-selector__price,
-	.wpcom-plan-selector__price-interval,
-	.wpcom-plan-selector__cta-label,
-	.wpcom-plan-selector__cta-component {
-		@include placeholder( --color-neutral-10 );
-		border-radius: 4px;
-		box-sizing: border-box;
-	}
-
-	.wpcom-plan-selector__owned-plan {
-		width: 150px;
-	}
-
-	.wpcom-plan-selector__plan-name {
-		height: 30px;
-		margin-block-end: 4px;
-		width: 200px;
-	}
-
-	.wpcom-plan-selector__price {
-		width: 260px;
-		margin-block-end: 2px;
-	}
-
-	.wpcom-plan-selector__price-interval {
-		width: 50px;
-		height: 14px;
-	}
-
-	.wpcom-plan-selector__cta {
-		gap: 16px;
-	}
-
-	.wpcom-plan-selector__cta-label {
-		width: 250px;
-	}
-
-	.wpcom-plan-selector__cta-component {
-		width: 436px;
-		height: 36px;
-	}
-}
