--- client/a8c-for-agencies/sections/marketplace/hosting-overview/hosting-v2/style.scss
@@ -1,274 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-
-$tab-background-color: #f4fbff;
-
-.hosting-v2 {
-
-	.a4a-slider__input {
-		margin-block-start: 0;
-	}
-
-	.components-button.pressable-overview-plan-selection__filter-button {
-		font-size: rem(12px);
-
-		@include break-mobile {
-			font-size: rem(14px);
-		}
-	}
-
-	.a4a-slider__input [type="range"] {
-		height: auto;
-		padding-block-end: 26px;
-	}
-
-	.a4a-slider__input-disabled-area {
-		inset-block-start: 0;
-	}
-	.a4a-slider__marker {
-		font-size: rem(10px);
-
-		@include break-mobile {
-			font-size: rem(12px);
-		}
-
-		@include break-medium {
-			font-size: rem(14px);
-		}
-	}
-
-	.a4a-slider__label-container {
-		padding-block-end: 16px;
-
-		@include break-medium {
-			padding-block-end: 0;
-		}
-	}
-
-	.hosting-v2__hero-container {
-		position: relative;
-		background-color: var(--color-primary-0);
-		border-block-end: 1px solid var(--color-accent-5);
-	}
-
-	.hosting-v2__hero {
-		position: relative;
-		overflow: hidden;
-		inset-block-start: 1px;
-
-		&::before {
-			background: url(calypso/assets/images/a8c-for-agencies/hosting-illustration.svg) no-repeat center/contain;
-			content: "";
-			display: inline-block;
-			position: absolute;
-			right: -150px;
-			top: -100px;
-			transition: transform 0.6s ease-in-out;
-			width: 200px;
-			height: 200px;
-
-			@include break-mobile {
-				right: -50px;
-			}
-
-			@include break-medium {
-				top: 0;
-			}
-
-			@include break-large {
-				right: 0;
-				top: -100px;
-			}
-
-			@include break-xlarge {
-				width: 250px;
-				height: 250px;
-			}
-
-			@include break-wide {
-				width: 300px;
-				height: 300px;
-			}
-
-			@include break-huge {
-				top: -100px;
-				width: 350px;
-				height: 350px;
-			}
-		}
-
-		&.is-standard {
-			&::before {
-				transform: scale(2.2) rotate(-120deg);
-			}
-		}
-
-		&.is-premier {
-			&::before {
-				transform: scale(2.2) rotate(-10deg);
-			}
-		}
-
-		&.is-enterprise {
-			&::before {
-				transform: scale(2.2) rotate(80deg);
-			}
-		}
-
-
-		padding: 0;
-
-		> * {
-			position: relative;
-			z-index: 1;
-			padding-inline: 16px;
-
-			@include break-large {
-				max-width: 1500px;
-				margin-inline: auto !important;
-				padding-inline: 40px;
-			}
-		}
-
-		.hosting-v2__hero-content {
-			padding: 16px 16px 0;
-
-			@include break-medium {
-				max-width: 1500px;
-				padding-inline: 40px;
-				padding-block-start: 32px;
-			}
-		}
-
-		.hosting-v2__container {
-			padding-block-start: 24px;
-
-			@include break-medium {
-				padding-block-start: 32px;
-			}
-		}
-
-		.hosting-v2__heading {
-			@include a4a-font-heading-xl;
-			margin-block-end: 24px;
-
-			@include break-medium {
-				@include a4a-font-heading-xxl;
-				margin-block-end: 32px;
-			}
-		}
-
-		.section-nav-tabs {
-			width: auto;
-			box-shadow: none;
-			margin: 24px 0 0;
-
-			@include break-medium {
-				margin: 39px 0 0;
-			}
-		}
-
-		.section-nav-tabs__list {
-			justify-content: space-between;
-			gap: 4px;
-		}
-
-		.section-nav-tab {
-			flex: 1;
-			border-radius: 4px 4px 2px 2px;
-			border: 1px solid var(--color-primary-5);
-			background: $tab-background-color;
-
-			&:hover:not(.is-selected) {
-				border-color: var(--color-primary-5);
-			}
-
-			&.is-selected {
-				border-color: var(--color-accent-5);
-				border-block-end: none;
-				background: var(--color-light-backdrop);
-
-				.section-nav-tab__link,
-				.section-nav-tab__link:hover {
-					background: unset;
-				}
-			}
-
-			.section-nav-tab__link:hover {
-				background-color: $tab-background-color;
-			}
-		}
-
-		.hosting-v2__nav-item-label {
-			color: var(--color-accent-90);
-			@include a4a-font-body-sm($font-weight: 600);
-
-			@include break-mobile {
-				@include a4a-font-heading-md($font-weight: 600);
-			}
-		}
-
-		.hosting-v2__nav-item-subtitle {
-			color: var(--color-accent-90);
-			@include a4a-font-body-md;
-		}
-
-	}
-
-	.hosting-v2__content-header {
-		display: none;
-
-		@include break-medium {
-			display: flex;
-			flex-direction: column;
-			align-items: center;
-			text-align: center;
-			gap: 16px;
-			padding: 0 40px 32px;
-
-			.hosting-v2__content-header-title {
-				@include a4a-font-body-md;
-			}
-
-			.hosting-v2__content-logo {
-				.wordpress-vip-logo {
-					display: inline-flex;
-				}
-
-				img {
-					height: 32px;
-				}
-			}
-		}
-	}
-
-	.hosting-v2__content {
-		padding-block-start: 32px;
-		position: relative;
-	}
-
-	.hosting-overview-features__item {
-		align-items: flex-start;
-		gap: 4px;
-
-		@include break-medium {
-			gap: 16px;
-		}
-	}
-
-	.hosting-overview-features__item-icon {
-		min-width: 32px;
-		max-width: 32px;
-		height: 32px;
-	}
-
-	.section-nav-tab__link {
-		padding: 16px 8px;
-		text-align: center;
-
-		@include break-mobile {
-			padding: 16px;
-		}
-	}
-}
-
