--- client/a8c-for-agencies/sections/marketplace/hosting-overview/style.scss
@@ -1,11 +1,0 @@
-.hosting-overview {
-	.hosting-dashboard-layout__body.is-full-width {
-		padding: 0;
-
-		& > * {
-			max-width: 100%;
-			padding: 0;
-			margin: 0;
-		}
-	}
-}
